Move necessary methods back down to MiqProvisionWorkflow
