+++ vmdb/app/models/miq_provision_virt_workflow.rb
@@ -4,18 +4,8 @@ class MiqProvisionVirtWorkflow < MiqProvisionWorkflow
     MiqProvisionInfraWorkflow
   }
 
-  def self.class_for_platform(platform)
-    "MiqProvision#{platform.titleize}Workflow".constantize
-  end
-
   def self.encrypted_options_fields
     [:root_password]
-  end
-
-  def self.class_for_source(source_or_id)
-    source = source_or_id.kind_of?(ActiveRecord) ? source_or_id : VmOrTemplate.find_by_id(source_or_id)
-    return nil if source.nil?
-    class_for_platform(source.class.model_suffix)
   end
 
   def self.request_class
@@ -859,28 +849,6 @@ class MiqProvisionVirtWorkflow < MiqProvisionWorkflow
     end
 
     return result
-  end
-
-  def self.validate_sysprep_file(io_handle)
-    require 'inifile'
-
-    begin
-      text = io_handle.read
-      if text.include?("<?xml")
-        xml = MiqXml.load(text)
-        raise "Invalid file contents detected" if xml.root.name != "unattend"
-      else
-        Tempfile.open('miqini') do |tf|
-          tf.write(text)
-          tf.close()
-          IniFile.load(tf.path)
-        end
-      end
-      return text
-    rescue StandardError, IniFile::Error
-      raise "Invalid file contents detected"
-    end
-    return nil
   end
 
   def allowed_number_of_vms(options={})
